Cr feedback - move SampleState to models.ts file - use SAMPLE_MANAGER_APP_PROPERTIES.controllerName - ManageSampleStatusesPanel.tsx add Cancel button for add new case
